redo structure of files
