+++ fetch_content.py
